Updated incoming request handler to pass header values as the same type, fixed a wasi-messaging bug, moved `orm` inside the `omnia-sdk`, and updated some dependencies.
