+++ crates/omnia-sdk/Cargo.toml
@@ -17,11 +17,13 @@ workspace = true
 anyhow.workspace = true
 axum = { workspace = true, features = ["json", "macros", "query"] }
 bytes.workspace = true
+chrono.workspace = true
 http.workspace = true
 http-body.workspace = true
 omnia-wasi-sql.workspace = true
 omnia-wasi-http.workspace = true
 omnia-wasi-jsondb.workspace = true
+sea-query.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror = "2.0.18"
