--- crates/orm/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "omnia-orm"
-description = "Object-relational mapping helper, suitable for use with omnia-wasi-sql"
-readme = "README.md"
-authors.workspace = true
-categories.workspace = true
-edition.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow.workspace = true
-chrono.workspace = true
-omnia-wasi-sql.workspace = true
-sea-query.workspace = true
-serde_json.workspace = true
-tracing.workspace = true
