+++ examples/sql/guest.rs
@@ -23,10 +23,10 @@ use axum::extract::Path;
 use axum::routing::{delete, get};
 use axum::{Json, Router};
 use chrono::Utc;
-use omnia_orm::{
-    DeleteBuilder, Entity, Filter, InsertBuilder, Join, SelectBuilder, UpdateBuilder, entity,
+use omnia_sdk::orm::{
+    DeleteBuilder, Entity, Filter, InsertBuilder, Join, SelectBuilder, UpdateBuilder,
 };
-use omnia_sdk::{HttpResult, TableStore};
+use omnia_sdk::{HttpResult, TableStore, entity};
 use omnia_wasi_sql::readwrite;
 use omnia_wasi_sql::types::{Connection, Statement};
 use serde::{Deserialize, Serialize};
